chore: Archive demo app in Makefile
This commit adds a new rule to the Makefile to archive the demo app. The `archive-demo-app` rule is responsible for creating a zip archive of the `demo-app` directory, excluding the `node_modules` directory. The resulting archive is then moved to the specified destination directory. This addition improves the project's build process by providing a convenient way to package and distribute the demo app.