refactor: Use the docker images of the web-app and demo-app by default
To have our dev env look as much as possible like the production
one. One working with the demo can still use --dev-demo to go back
to the live one.

Also, in order to ease moving from one image to the other, all the listened
ports are now the same between web/demo app and docker/live images. It is now
4200 because, well, why not!?