doc: API documentation updates
The code changes include adding a new type `KeybanApiStatus` to represent the status of the Keyban API. This type is added to the `api.ts` file in the `sdk-base` package. Additionally, the `api.tsx` file in the `sdk-react` package is updated to include a new function `useKeybanApiStatus` that utilizes the `KeybanApiStatus` type.