chore: Update npm dependencies and add @keyban/mui-theme package to the loaylty demo
The commit updates the npm dependencies in the `package.json` file of the `loyalty-demo` app. It adds the `@keyban/mui-theme` package as a dependency. This package is required to apply the Keyban theme to the app. The version specified is `workspace:*`, indicating that the latest version from the workspace should be used.