feat: Add name column to applications table
This commit adds a new column named "name" to the applications table in the database. The column is of type TEXT and has a default value of 'app_' followed by a randomly generated UUID. The commit also includes an update to set the name of a specific application to 'waas-demo' and an insert statement to add a new application with a specific app_id, jwks_uri, and name.