refactor: installation and setup documentation
The installation and setup documentation for the wallet-as-a-service has been removed as it is no longer needed. This will help streamline the documentation and remove any confusion for users.

Refactor the sidebars.ts file

The sidebars.ts file has been refactored to remove the installation and setup section from the sidebar. This change aligns with the removal of the installation and setup documentation.

Update package descriptions for Keyban SDK Base and Keyban SDK React

The package descriptions for Keyban SDK Base and Keyban SDK React have been updated. This will provide more accurate and informative descriptions for these packages.

close #620